03-04-25

+++ src/corlib/Internal/Runtime/CompilerHelpers/RuntimeImports.cs
@@ -22,7 +22,7 @@ namespace Internal.Runtime.CompilerHelpers
 
             var b = (byte*)data;
             b += sizeof(IntPtr);
-            MemoryHelpers.MemCpy(b, (byte*)&length, sizeof(long));
+            MemoryHelpers.MemCpy(b, (byte*)&length, sizeof(int));
 
             return obj;
         }
